refactor(tauri)!: runtime is a pure derivation, delete IRuntime (PR-4 T7)
Legacy writable runtime dies here: Config::runtime()/IRuntime,
generate_file()/runtime_config_path()/ConfigType, and the
RUNTIME_CONFIG_DIR/RUNTIME_CONFIG/CHECK_CONFIG constants in
config/core.rs are deleted (config/core.rs now holds only the
clash/verge domains). draft_define!(IRuntime) and the IRuntime import
are removed from draft.rs; config/mod.rs drops the runtime module.

client/mod.rs: regenerate_runtime_with no longer dual-writes a legacy
IRuntime into Config::runtime().draft() after publishing to the
RuntimeState manager (spawn_blocking now returns (RuntimeState,
String)); the "runtime draft written" legacy test assertion is
removed, keeping the Task 2 manager assertions.

enhance/artifact_bridge.rs: the runtime_from_artifact() thin shell
over runtime_state_from_artifact is deleted along with its IRuntime
import and export; the module doc is corrected to describe the
RuntimeState read model instead of "the legacy IRuntime triple".

PatchRuntimeConfig moves verbatim into ipc.rs, directly above
patch_clash_config; wire format is unchanged.

client/runtime.rs: two doc comments (RuntimeState's header, and
runtime_config_path's doc) referenced the now-deleted IRuntime /
Config::runtime_config_path() by name for historical context; reworded
(no logic change) so backend/tauri/src is grep-clean for
Config::runtime|IRuntime|generate_file|ConfigType per the task
postcondition.

Postcondition verified: rg for
Config::runtime|IRuntime|generate_file|ConfigType across
backend/tauri/src returns zero hits. cargo test --workspace
--all-features passes (234/234 in the tauri lib crate; a single
bridge::verge:: failure was observed once and reproduced as
pre-existing parallel-test flakiness from the shared Config::global()
singleton, unrelated to this change — confirmed by an isolated rerun
and two subsequent clean full-suite runs).
