+++ backend/tauri/src/client/mod.rs
@@ -725,12 +725,8 @@ impl NyanpasuClient {
         let profiles_dir = self.inner.profiles_dir.clone();
         let core = app.core;
         let builtin_enabled = app.enable_builtin_enhanced;
-        let (state, legacy_runtime, yaml) = tokio::task::spawn_blocking(
-            move || -> anyhow::Result<(
-                crate::client::runtime::RuntimeState,
-                crate::config::IRuntime,
-                String,
-            )> {
+        let (state, yaml) = tokio::task::spawn_blocking(
+            move || -> anyhow::Result<(crate::client::runtime::RuntimeState, String)> {
                 let content = FsProfileContentSource::new(profiles_dir);
                 let scripts = EnhanceScriptRunner::new()?;
                 let input = RuntimeBuildInput {
@@ -742,16 +738,11 @@ impl NyanpasuClient {
                 let artifact = RuntimeBuilder::build(&input, &content, &scripts)?;
                 let state =
                     runtime_state_from_artifact(&artifact, &profiles, core, builtin_enabled)?;
-                let legacy_runtime = crate::config::IRuntime {
-                    config: Some(state.config.clone()),
-                    exists_keys: state.exists_keys.clone(),
-                    postprocessing_output: state.postprocessing_output.clone(),
-                };
                 let yaml = format!(
                     "# Generated by Clash Nyanpasu\n\n{}",
                     serde_yaml::to_string(&state.config)?
                 );
-                Ok((state, legacy_runtime, yaml))
+                Ok((state, yaml))
             },
         )
         .await
@@ -776,9 +767,6 @@ impl NyanpasuClient {
                 ClientError::Custom(format!("failed to store runtime state: {error}"))
             })?;
         }
-        // TODO(actor-migration): temporary dual-write to Config::runtime() draft (B8).
-        // Reason: legacy readers migrate in PR-4 Tasks 3-6. Remove in PR-4 Task 7.
-        *crate::config::Config::runtime().draft() = legacy_runtime;
         Ok(())
     }
 }
@@ -1208,10 +1196,6 @@ mod tests {
                 .activate_profile(Some(uid.clone()))
                 .await
                 .expect("activate");
-            let runtime = crate::config::Config::runtime();
-            let runtime = runtime.latest();
-            let config = runtime.config.as_ref().expect("runtime draft written");
-            assert!(config.get("mixed-port").is_some());
             let state = client.runtime_state().await;
             let state = state
                 .as_ref()
