+++ backend/tauri/src/config/core.rs
@@ -1,17 +1,9 @@
-use super::{Draft, IClashTemp, IRuntime, IVerge};
-use crate::utils::{dirs, help};
-use anyhow::{Result, anyhow};
+use super::{Draft, IClashTemp, IVerge};
 use once_cell::sync::OnceCell;
-use std::{env::temp_dir, path::PathBuf};
-
-pub const RUNTIME_CONFIG_DIR: &str = "runtime";
-pub const RUNTIME_CONFIG: &str = "clash-config.yaml";
-pub const CHECK_CONFIG: &str = "clash-config-check.yaml";
 
 pub struct Config {
     clash_config: Draft<IClashTemp>,
     verge_config: Draft<IVerge>,
-    runtime_config: Draft<IRuntime>,
 }
 
 impl Config {
@@ -21,7 +13,6 @@ impl Config {
         CONFIG.get_or_init(|| Config {
             clash_config: Draft::from(IClashTemp::new()),
             verge_config: Draft::from(IVerge::new()),
-            runtime_config: Draft::from(IRuntime::new()),
         })
     }
 
@@ -32,41 +23,4 @@ impl Config {
     pub fn verge() -> Draft<IVerge> {
         Self::global().verge_config.clone()
     }
-
-    pub fn runtime() -> Draft<IRuntime> {
-        Self::global().runtime_config.clone()
-    }
-
-    /// 将配置丢到对应的文件中
-    pub fn generate_file(typ: ConfigType) -> Result<PathBuf> {
-        let path = match typ {
-            ConfigType::Run => Self::runtime_config_path()?,
-            ConfigType::Check => temp_dir().join(CHECK_CONFIG),
-        };
-        if let Some(parent) = path.parent() {
-            std::fs::create_dir_all(parent)?;
-        }
-
-        let runtime = Config::runtime();
-        let runtime = runtime.latest();
-        let config = runtime
-            .config
-            .as_ref()
-            .ok_or(anyhow!("failed to get runtime config"))?;
-
-        help::save_yaml(&path, &config, Some("# Generated by Clash Nyanpasu"))?;
-        Ok(path)
-    }
-
-    pub(crate) fn runtime_config_path() -> Result<PathBuf> {
-        Ok(dirs::app_config_dir()?
-            .join(RUNTIME_CONFIG_DIR)
-            .join(RUNTIME_CONFIG))
-    }
-}
-
-#[derive(Debug)]
-pub enum ConfigType {
-    Run,
-    Check,
 }
