+++ backend/tauri/src/config/mod.rs
@@ -2,7 +2,6 @@ mod clash;
 mod core;
 mod draft;
 pub mod nyanpasu;
-mod runtime;
-pub use self::{clash::*, core::*, draft::*, runtime::*};
+pub use self::{clash::*, core::*, draft::*};
 
 pub use self::nyanpasu::IVerge;
