+++ backend/tauri/src/enhance/mod.rs
@@ -10,7 +10,7 @@ mod golden;
 #[cfg(test)]
 pub(crate) mod golden_support;
 
-pub use artifact_bridge::{runtime_from_artifact, runtime_state_from_artifact};
+pub use artifact_bridge::runtime_state_from_artifact;
 pub use content_source::FsProfileContentSource;
 pub use runtime_builder::{
     RuntimeBuildError, RuntimeBuildInput, RuntimeBuilder, builtin_transforms_for, derive_tun_flavor,
